@@ -3,17 +3,14 @@ plugins {
 }
 
 android {
-    compileSdk 31
+    compileSdk 33
     defaultConfig {
         minSdk 21
-        targetSdk 31
+        targetSdk 33
     }
 
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
+    buildTypes.all {
+        minifyEnabled true
     }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
@@ -22,5 +19,5 @@ android {
 }
 
 dependencies {
-    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'androidx.appcompat:appcompat:1.5.1'
 }
