@@ -1,14 +1,14 @@
 package com.alex.shimmer;
 
 public interface ShimmerViewBase {
-    public float getGradientX();
-    public void setGradientX(float gradientX);
-    public boolean isShimmering();
-    public void setShimmering(boolean isShimmering);
-    public boolean isSetUp();
-    public void setAnimationSetupCallback(ShimmerViewHelper.AnimationSetupCallback callback);
-    public int getPrimaryColor();
-    public void setPrimaryColor(int primaryColor);
-    public int getReflectionColor();
-    public void setReflectionColor(int reflectionColor);
+    float getGradientX();
+    void setGradientX(float gradientX);
+    boolean isShimmering();
+    void setShimmering(boolean isShimmering);
+    boolean isSetUp();
+    void setAnimationSetupCallback(ShimmerViewHelper.AnimationSetupCallback callback);
+    int getPrimaryColor();
+    void setPrimaryColor(int primaryColor);
+    int getReflectionColor();
+    void setReflectionColor(int reflectionColor);
 }
